[ci skip] Rename pkg -> solr
- the test would break for sure ...
