--- pkg/util/log.go
@@ -1,18 +1,0 @@
-package util
-
-import (
-	dlog "github.com/dyweb/gommon/log"
-)
-
-// Logger is the default logger with info level
-var Logger = dlog.NewLogger()
-
-// Short name use in util package
-var log = Logger.RegisterPkg()
-
-func init() {
-	f := dlog.NewTextFormatter()
-	f.EnableColor = true
-	Logger.Formatter = f
-	Logger.Level = dlog.InfoLevel
-}
