--- pkg/util/test.go
@@ -1,1 +1,0 @@
-package util
