+++ solr/admin/core.go
