+++ solr/admin/core_test.go
