+++ solr/admin/info.go
