+++ solr/admin/pkg.go
