+++ solr/admin/pkg_test.go
