+++ solr/common/constant.go
