+++ solr/common/entity.go
