+++ solr/common/entity_test.go
