+++ solr/common/error.go
