+++ solr/common/fieldtype/pkg.go
