+++ solr/common/response.go
