+++ solr/core/admin.go
