+++ solr/core/admin_test.go
