+++ solr/core/fixture/facet.json
