+++ solr/core/pkg.go
