+++ solr/core/pkg_test.go
