+++ solr/core/select.go
