+++ solr/core/select_test.go
