+++ solr/core/update.go
