+++ solr/core/update_test.go
