+++ solr/fixture/schema.go
