+++ solr/internal/client.go
