+++ solr/internal/client_test.go
