+++ solr/internal/pkg.go
