+++ solr/internal/test_util.go
