+++ solr/schema/field.go
