+++ solr/schema/filed_test.go
