+++ solr/schema/infer.go
