+++ solr/schema/infer_test.go
