+++ solr/schema/pkg.go
