+++ solr/schema/pkg_test.go
