+++ solr/search/pkg.go
