+++ solr/search/query.go
