+++ solr/search/query_test.go
